# Optimize loader.