Hi @scsosna99, 

my friend and colleague @jexp brought this post https://dzone.com/articles/improving-neo4j-ogm-performance to our attention.

Your analysis of the `Neo4jSession` is very detailed and pretty accurate, but leads the reader to thin ice, especially when the reader decides that manipulating the internals is a good idea.

With this pull request I'd like to propose a better solution:

The main idea around the session is: It is a *short lived object* That is: You don't keep it around. When it's used with Spring Data Neo4j, for each transaction a new one is created and that's what we do with the PR.

We keep the `SessionFactory` ( which is expensive to create) and than open sessions as needed. Especially in the `processFilings` methods.
We do the same inside the cache population.

On top of that, I have upgrade to just released version of Neo4j-OGM 3.2.1.
For the creating the indexes and purging the database, I find it much better to `unwrap` the underlying java driver as shown.

As I'm writing this, my 13" MacBook Pro is printing this output:

```
Processing 2015_3_7_17.xml: 716 filings in 2432 ms
Processing 2015_3_7_2.xml: 1000 filings in 4053 ms
Processing 2015_3_7_3.xml: 1000 filings in 4187 ms
Processing 2015_3_7_4.xml: 1000 filings in 6272 ms
Processing 2015_3_7_5.xml: 1000 filings in 7010 ms
Processing 2015_3_7_6.xml: 1000 filings in 5644 ms
Processing 2015_3_7_7.xml: 1000 filings in 6520 ms
Processing 2015_3_7_8.xml: 1000 filings in 5534 ms
Processing 2015_3_7_9.xml: 1000 filings in 4783 ms
Processing 2015_3_8_1.xml: 1000 filings in 2704 ms
Processing 2015_3_8_2.xml: 226 filings in 620 ms
Processing 2015_3_9_1.xml: 703 filings in 1702 ms
Processing 2015_4_10_1.xml: 1000 filings in 3988 ms
```

I really hope this PR made it's way into your project and in a potential update to your otherwise grate post on DZone.

Thanks,
Michael.