@michael-simons 2019-10-14
I have added some more improvements, mostly what Michael Hunger recommended (using natural keys where possible), but also some improvements with the caching and using native types.