@scsosna99 2019-10-15
Actually, you're confirming what I discovered: the cache handling in Session is garbage and needs some love and attention.  My experience/expectation is that caches increase, not decrease, performance, and definitely not as substantially as OGM's does.

The argument that a Session should be short-lived and re-creating for each and every Neo4J call - your initial commit - makes no sense, because why would there even be the need for the intenal caching if each call is a unique session.  Also, transactionality brings no value if each and every call requires a unique session.  Since opening resource connections - e.g., connections to databases, message servers, other resources - is often an extremely expensive operation, suggesting that this should be done approximately 3M times (for the 5 years of data, assuming the program caching is working right), is silly unless SessionFactory has some internal connection pooling (which I haven't investigated yet).

Your later change where each filing uses a session is better, but believe I tested this way and did not see a noticeable gain.  It was only when I cleared the session that performance increased, I'll have to retest at some point.

Though I did provide a method which uses reflection to do nasty stuff, the current version has calls to the method commented out and only calls Session.clear(), which is declared on the interface so no real nastiness.  It just confirms my contention that clearing the Session cache results in magnitudes-better performance.

The other changes made later in the day are minor with minimal impact, nothing on the scale of clearing the session each time through or opening up a new session each time.  The program's caches max sizes had relatively no impact on overall performance, whether 1, 100, 5000.  Changing cache implementation, ensuring the zip files are closed, clearing the database via Cypher than the Session call are all relatively uneventful.  It's the Session cache to focus on.

The main point for anyone using OGM to understand is that Session's internal caching of nodes and relationships is a major bottleneck on performance and a seemingly easy way to circumvent is to clear regularly and keep the size down - and doesn't seem to break anything.  I did mention that two successive files saw a dramatic increase when run in a single session, but also tested per-filing times which did go up (but didn't document).  Hopefully there's upcoming work on Session to address this.

At this point, I'm not accepting the pull request, I'll look at your changes individually and do some more research but don't believe that my basic argument is incorrect or misplaced.