@michael-simons 2019-10-15
Well, you’re missing the point. If you open a session factory 3m times, that does open connections etc. 
Think what you want, but we as the authors tell you the session is a short lived object. Opening costs nothing. You can either use it as intended to be, or do you’re thing.