@scsosna99 2019-10-16
I'll need to dig in deeper to the OGM code to understand better, changing to a per-filing basis is possible since it provides the transactionality that I want.  The big takeaway of my article was the performance hit caused by the cache, for those like me who assumed the DB connection is a heavy-weight object and should be reused or pooled or whatever one does for efficient resource management.

If you don't want people to manipulate the cache as I am, then perhaps that method should be removed from the interface.  Especially if your intention is a short-lived object, not sure what the use case for even having the cache is.  Again, might become clearing once I dig into the OGM code and start understanding its internals.

I'll get back to this when I can, but since I just started a new job it's going to be when time permits vs. the open-ended time I had this summer when I was between gigs.

Thanks for your feedback.